Nmap 2.01
git-svn-id: https://svn.nmap.org/trunk/nmap@465 e0a8ed71-7df4-0310-8962-fdc924857419

+++ targets.c
@@ -149,7 +149,7 @@ addy[0] = r = hostexp;
    + the optional '/mask' */
 target_net = strtok(hostexp, "/");
 targets->netmask = (int) (s = strtok(NULL,""))? atoi(s) : 32;
-if (targets->netmask < 0 || targets->netmask > 32) {
+if ((int) targets->netmask < 0 || targets->netmask > 32) {
   printf("Illegal netmask value (%d), must be /0 - /32 .  Assuming /32 (one host)\n", targets->netmask);
   targets->netmask = 32;
 }
